+++ docs/metrics/getting-started-prometheus-grafana/Program.cs
@@ -52,6 +52,6 @@ public class Program
 
         // Dispose meter provider before the application ends.
         // This will flush the remaining metrics and shutdown the metrics pipeline.
-        meterProvider?.Dispose();
+        meterProvider.Dispose();
     }
 }